Refactor EditTimelineFilterDialog
