+++ composeApp/src/desktopMain/kotlin/com/alekso/dltstudio/timeline/filters/TimelineFiltersDialog.kt
@@ -21,6 +21,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.DialogWindow
 import androidx.compose.ui.window.rememberDialogState
 import com.alekso.dltstudio.timeline.DiagramType
+import com.alekso.dltstudio.timeline.filters.edit.EditTimelineFilterDialog
+import com.alekso.dltstudio.timeline.filters.edit.EditTimelineFilterDialogState
 import com.alekso.dltstudio.ui.CustomButton
 import com.alekso.dltstudio.ui.CustomCheckbox
 import com.alekso.dltstudio.ui.ImageButton
@@ -51,8 +53,8 @@ fun TimelineFiltersDialog(
             EditTimelineFilterDialog(
                 visible = editDialogState.value.visible,
                 onDialogClosed = { editDialogState.value = EditTimelineFilterDialogState(false) },
-                colorFilter = editDialogState.value.filter,
-                colorFilterIndex = editDialogState.value.filterIndex,
+                timelineFilter = editDialogState.value.filter,
+                filterIndex = editDialogState.value.filterIndex,
                 onFilterUpdate = { i, filter ->
                     editDialogState.value.filter = filter
                     onTimelineFilterUpdate(i, filter)
